reduce index usage for scanning arrow batches

+++ include/lingodb/runtime/ArrowView.h
@@ -6,15 +6,15 @@
 
 namespace lingodb::runtime {
 struct ArrayView {
-   static constexpr int64_t maxNullCount = 1 << 20;
-   static constexpr int64_t validDataLength = maxNullCount / 8;
+   static constexpr int32_t maxNullCount = 1 << 20;
+   static constexpr int32_t validDataLength = maxNullCount / 8;
    static std::array<uint8_t, validDataLength> validData;
    // Array data description
-   int64_t length;
-   int64_t nullCount;
-   int64_t offset;
-   int64_t nBuffers;
-   int64_t nChildren;
+   int32_t length;
+   int32_t nullCount;
+   int32_t offset;
+   int32_t nBuffers;
+   int32_t nChildren;
    const void** buffers;
    const ArrayView** children;
 };
@@ -22,8 +22,8 @@ struct ArrayView {
 struct BatchView {
    static std::array<uint16_t, 65536> defaultSelectionVector;
    static constexpr size_t maxBatchSize = 65536;
-   int64_t length;
-   int64_t offset;
+   int32_t length;
+   int32_t offset;
    uint16_t* selectionVector;
    const ArrayView** arrays;
 };
